+++ CHANGES.md
@@ -8,6 +8,7 @@ CHANGES
 * Added `arePendingFriends` and `isPendingFriend `to Model and Document Classes
 * updated tests for above additions
 * `acceptRequest` and `denyRequest` now accept two accounts `_ids` in either order instead of requester then requested
+* added `isRequester` and `isRequested` to `AccountModel` and `AccountDocument`
 
 ## v1.1.0
 * `lib/relationships.js` now includes relationship state `PENDING_FRIENDS` with value `1.5` for backward compatibility